requirements.txt
grpcio==1.62.1
grpcio-tools==1.62.1
protobuf==4.25.3
pydantic==2.6.4
pydantic-settings==2.2.1
graia-broadcast==0.23.5
graia-amnesia==0.9.0
launart==0.8.2